+++ src/main/kotlin/com/iwayee/activity/api/comp/Member.kt
@@ -4,11 +4,11 @@ import com.iwayee.activity.define.GroupPosition
 import com.iwayee.activity.define.SexType
 
 data class Member(
-        var id: Int = 0,
+        var id: Long = 0,
         var scores: Int = 0,
         var pos: Int = GroupPosition.POS_MEMBER.ordinal,
         var sex: Int = SexType.MALE.ordinal,
-        var at: Long = 0L,
+        var at: Long = 0,
         var nick: String = "",
         var wx_nick: String = ""
 ) {
