+++ src/main/kotlin/com/iwayee/activity/api/comp/Session.kt
@@ -3,8 +3,8 @@ package com.iwayee.activity.api.comp
 import com.iwayee.activity.define.SexType
 
 data class Session(
-        var uid: Int = 0,
+        var uid: Long = 0,
         var sex: Int = SexType.MALE.ordinal,
-        var at: Long = 0L,
+        var at: Long = 0,
         var token: String = ""
 )
